TP8: Q8.5

+++ tp-spring/src/main/java/web/controller/TodoControllerV2.java
@@ -9,6 +9,7 @@ import web.model.Category;
 import web.model.Todo;
 import web.service.TodoService;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 
@@ -26,31 +27,42 @@ public class TodoControllerV2 {
     }
 
     @GetMapping(path = "todo", produces = MediaType.APPLICATION_JSON_VALUE)
-    public Todo get() {
-        return new Todo(1, "A title", "desc", List.of(Category.ENTERTAINMENT, Category.WORK), null, "foo");
+    public Todo get(Principal user) {
+        return new Todo(1, "A title", "desc", List.of(Category.ENTERTAINMENT, Category.WORK), null, user.getName());
     }
 
     @PostMapping(path = "todo",
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Todo post(@RequestBody final Todo todo) {
+        consumes = MediaType.APPLICATION_JSON_VALUE,
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    public Todo post(final Principal user, @RequestBody final Todo todo) {
+        todo.setId(0); // ensure not to overwrite an existing todo
+        todo.setOwner(user.getName()); // ensure the owner is the logged user
         return todoService.add(todo);
     }
 
     @PutMapping(path = "todo",
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Todo put(@RequestBody final Todo todo) {
+        consumes = MediaType.APPLICATION_JSON_VALUE,
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    public Todo put(final Principal user, @RequestBody final Todo todo) {
+        // Ensure the todo exists and is owned by the logger user
+        if (!todoService.existByIdAndOwner(todo.getId(), user.getName()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        todo.setOwner(user.getName()); // ensure the owner is not changed
         if (!todoService.replace(todo))
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         return todo;
     }
 
     @PatchMapping(path = "todo/{id}",
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Todo patch(@PathVariable final long id,
+        consumes = MediaType.APPLICATION_JSON_VALUE,
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    public Todo patch(final Principal user, @PathVariable final long id,
                       @RequestBody final Map<String, Object> partialTodo) {
+        // Ensure the todo exists and is owned by the logger user
+        if (!todoService.existByIdAndOwner(id, user.getName()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        partialTodo.remove("id"); // ensure the id is not modified
+        partialTodo.remove("owner"); // ensure the owner is not modified
         Todo modifiedTodo = todoService.modify(id, partialTodo);
         if (modifiedTodo == null)
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
@@ -58,7 +70,10 @@ public class TodoControllerV2 {
     }
 
     @DeleteMapping(path = "todo/{id}")
-    public void delete(@PathVariable final long id) {
+    public void delete(final Principal user, @PathVariable final long id) {
+        // Ensure the todo exists and is owned by the logger user
+        if (!todoService.existByIdAndOwner(id, user.getName()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         if (!todoService.delete(id))
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
     }
